Handle Relationship User and Contact Schema, Protecting Routes(contact), Logged in User CRUD operations

+++ controllers/contactController.js
@@ -5,23 +5,28 @@ const Contact = require("../models/contactModel");
 
 // @desc Get all contacts
 // @route GET /api/contacts
-// @access public
+// @access private
 const getContacts = asyncHandler(async (req, res, next) => {
-  const contacts = await Contact.find();
+  const contacts = await Contact.find({ user_id: req.user.id });
   // res.status(200).json({ message: "Get all contacts" });
   res.status(200).json(contacts);
 });
 
 // @desc Create New contact
 // @route POST /api/contacts
-// @access public
+// @access private
 const createContact = asyncHandler(async (req, res, next) => {
   const { name, email, phone } = req.body;
   if (!name || !email || !phone) {
     res.status(400);
     throw new Error("all fields are compulsory");
   }
-  const contact = await Contact.create({ name, email, phone });
+  const contact = await Contact.create({
+    name,
+    email,
+    phone,
+    user_id: req.user.id,
+  });
   // res.status(201).json({ message: "Create contact" });
   res.status(201).json(contact);
   // 201 --> resource created
@@ -29,7 +34,7 @@ const createContact = asyncHandler(async (req, res, next) => {
 
 // @desc Get contact
 // @route GET /api/contacts/:id
-// @access public
+// @access private
 const getContact = asyncHandler(async (req, res, next) => {
   const contact = await Contact.findById(req.params.id);
   if (!contact) {
@@ -42,13 +47,20 @@ const getContact = asyncHandler(async (req, res, next) => {
 
 // @desc Update contact
 // @route PUT /api/contacts/:id
-// @access public
+// @access private
 const updateContact = asyncHandler(async (req, res, next) => {
   const contact = await Contact.findById(req.params.id);
   if (!contact) {
     res.status(404);
     throw new Error("Contact Not Found");
   }
+
+  // stopping different user from updating contacts of some other person
+  if (contact.user_id.toString() !== req.user.id) {
+    res.status(403);
+    throw new Error("User don't have permission to update other user contacts");
+  }
+
   const updatedContact = await Contact.findByIdAndUpdate(
     req.params.id,
     req.body,
@@ -63,7 +75,20 @@ const updateContact = asyncHandler(async (req, res, next) => {
 // @route DELETE /api/contacts/:id
 // @access public
 const deleteContact = asyncHandler(async (req, res, next) => {
-  const contact = await Contact.findByIdAndDelete(req.params.id);
+  let contact = await Contact.findById(req.params.id);
+  if (!contact) {
+    res.status(404);
+    throw new Error("Contact Not Found");
+  }
+
+  // stopping different user from deleting contacts of some other person
+  if (contact.user_id.toString() !== req.user.id) {
+    res.status(403);
+    throw new Error("User don't have permission to delete other user contacts");
+  }
+
+  // removing contact
+  contact = await Contact.findByIdAndDelete(req.params.id);
   res.status(200).json(contact);
 
   // res.status(201).json({ message: `Delete contact for ${req.params.id}` });
